+++ lib/spack/spack/test/ci.py
@@ -13,8 +13,12 @@ import spack.ci as ci
 import spack.environment as ev
 import spack.error
 import spack.paths as spack_paths
+import spack.repo as repo
 import spack.spec
 import spack.util.git
+from spack.spec import Spec
+
+pytestmark = [pytest.mark.usefixtures("mock_packages")]
 
 
 @pytest.fixture
@@ -25,8 +29,128 @@ def repro_dir(tmp_path):
         yield result
 
 
-def test_urlencode_string():
-    assert ci._url_encode_string("Spack Test Project") == "Spack+Test+Project"
+def test_pipeline_dag(config, tmpdir):
+    r"""Test creation, pruning, and traversal of PipelineDAG using the
+    following package dependency graph:
+
+        a                           a
+       /|                          /|
+      c b                         c b
+        |\        prune 'd'        /|\
+        e d        =====>         e | g
+        | |\                      | |
+        h | g                     h |
+         \|                        \|
+          f                         f
+
+    """
+    builder = repo.MockRepositoryBuilder(tmpdir)
+    builder.add_package("pkg-h", dependencies=[("pkg-f", None, None)])
+    builder.add_package("pkg-g")
+    builder.add_package("pkg-f")
+    builder.add_package("pkg-e", dependencies=[("pkg-h", None, None)])
+    builder.add_package("pkg-d", dependencies=[("pkg-f", None, None), ("pkg-g", None, None)])
+    builder.add_package("pkg-c")
+    builder.add_package("pkg-b", dependencies=[("pkg-d", None, None), ("pkg-e", None, None)])
+    builder.add_package("pkg-a", dependencies=[("pkg-b", None, None), ("pkg-c", None, None)])
+
+    with repo.use_repositories(builder.root):
+        spec_a = Spec("pkg-a").concretized()
+
+        key_a = ci.common.PipelineDag.key(spec_a)
+        key_b = ci.common.PipelineDag.key(spec_a["pkg-b"])
+        key_c = ci.common.PipelineDag.key(spec_a["pkg-c"])
+        key_d = ci.common.PipelineDag.key(spec_a["pkg-d"])
+        key_e = ci.common.PipelineDag.key(spec_a["pkg-e"])
+        key_f = ci.common.PipelineDag.key(spec_a["pkg-f"])
+        key_g = ci.common.PipelineDag.key(spec_a["pkg-g"])
+        key_h = ci.common.PipelineDag.key(spec_a["pkg-h"])
+
+        pipeline = ci.common.PipelineDag([spec_a])
+
+        expected_bottom_up_traversal = {
+            key_a: 4,
+            key_b: 3,
+            key_c: 0,
+            key_d: 1,
+            key_e: 2,
+            key_f: 0,
+            key_g: 0,
+            key_h: 1,
+        }
+
+        visited = []
+        for stage, node in pipeline.traverse_nodes(direction="parents"):
+            assert expected_bottom_up_traversal[node.key] == stage
+            visited.append(node.key)
+
+        assert len(visited) == len(expected_bottom_up_traversal)
+        assert all(k in visited for k in expected_bottom_up_traversal.keys())
+
+        expected_top_down_traversal = {
+            key_a: 0,
+            key_b: 1,
+            key_c: 1,
+            key_d: 2,
+            key_e: 2,
+            key_f: 4,
+            key_g: 3,
+            key_h: 3,
+        }
+
+        visited = []
+        for stage, node in pipeline.traverse_nodes(direction="children"):
+            assert expected_top_down_traversal[node.key] == stage
+            visited.append(node.key)
+
+        assert len(visited) == len(expected_top_down_traversal)
+        assert all(k in visited for k in expected_top_down_traversal.keys())
+
+        pipeline.prune(key_d)
+        b_children = pipeline.nodes[key_b].children
+        assert len(b_children) == 3
+        assert all([k in b_children for k in [key_e, key_f, key_g]])
+
+        # check another bottom-up traversal after pruning pkg-d
+        expected_bottom_up_traversal = {
+            key_a: 4,
+            key_b: 3,
+            key_c: 0,
+            key_e: 2,
+            key_f: 0,
+            key_g: 0,
+            key_h: 1,
+        }
+
+        visited = []
+        for stage, node in pipeline.traverse_nodes(direction="parents"):
+            assert expected_bottom_up_traversal[node.key] == stage
+            visited.append(node.key)
+
+        assert len(visited) == len(expected_bottom_up_traversal)
+        assert all(k in visited for k in expected_bottom_up_traversal.keys())
+
+        # check top-down traversal after pruning pkg-d
+        expected_top_down_traversal = {
+            key_a: 0,
+            key_b: 1,
+            key_c: 1,
+            key_e: 2,
+            key_f: 4,
+            key_g: 2,
+            key_h: 3,
+        }
+
+        visited = []
+        for stage, node in pipeline.traverse_nodes(direction="children"):
+            assert expected_top_down_traversal[node.key] == stage
+            visited.append(node.key)
+
+        assert len(visited) == len(expected_top_down_traversal)
+        assert all(k in visited for k in expected_top_down_traversal.keys())
+
+        a_deps_direct = [n.spec for n in pipeline.get_dependencies(pipeline.nodes[key_a])]
+        assert all([s in a_deps_direct for s in [spec_a["pkg-b"], spec_a["pkg-c"]]])
 
 
 @pytest.mark.not_on_windows("Not supported on Windows (yet)")
