+++ lib/spack/spack/test/cmd/ci.py
@@ -23,8 +23,10 @@ import spack.environment as ev
 import spack.hash_types as ht
 import spack.main
 import spack.paths as spack_paths
-import spack.repo as repo
 import spack.util.spack_yaml as syaml
+from spack.ci import gitlab as gitlab_generator
+from spack.ci.common import PipelineDag, PipelineOptions, SpackCIConfig
+from spack.ci.generator_registry import generator
 from spack.cmd.ci import FAILED_CREATE_BUILDCACHE_CODE
 from spack.schema.buildcache_spec import schema as specfile_schema
 from spack.schema.database_index import schema as db_idx_schema
@@ -117,62 +119,6 @@ def ci_generate_test(tmp_path, mutable_mock_env_path, install_mockery, ci_base_e
         return spack_yaml, outputfile, output
 
     return _func
-
-
-def test_specs_staging(config, tmpdir):
-    """Make sure we achieve the best possible staging for the following
-spec DAG::
-
-        a
-       /|
-      c b
-        |\
-        e d
-          |\
-          f g
-
-In this case, we would expect 'c', 'e', 'f', and 'g' to be in the first stage,
-and then 'd', 'b', and 'a' to be put in the next three stages, respectively.
-
-"""
-    builder = repo.MockRepositoryBuilder(tmpdir)
-    builder.add_package("pkg-g")
-    builder.add_package("pkg-f")
-    builder.add_package("pkg-e")
-    builder.add_package("pkg-d", dependencies=[("pkg-f", None, None), ("pkg-g", None, None)])
-    builder.add_package("pkg-c")
-    builder.add_package("pkg-b", dependencies=[("pkg-d", None, None), ("pkg-e", None, None)])
-    builder.add_package("pkg-a", dependencies=[("pkg-b", None, None), ("pkg-c", None, None)])
-
-    with repo.use_repositories(builder.root):
-        spec_a = Spec("pkg-a").concretized()
-
-        spec_a_label = ci._spec_ci_label(spec_a)
-        spec_b_label = ci._spec_ci_label(spec_a["pkg-b"])
-        spec_c_label = ci._spec_ci_label(spec_a["pkg-c"])
-        spec_d_label = ci._spec_ci_label(spec_a["pkg-d"])
-        spec_e_label = ci._spec_ci_label(spec_a["pkg-e"])
-        spec_f_label = ci._spec_ci_label(spec_a["pkg-f"])
-        spec_g_label = ci._spec_ci_label(spec_a["pkg-g"])
-
-        spec_labels, dependencies, stages = ci.stage_spec_jobs([spec_a])
-
-        assert len(stages) == 4
-
-        assert len(stages[0]) == 4
-        assert spec_c_label in stages[0]
-        assert spec_e_label in stages[0]
-        assert spec_f_label in stages[0]
-        assert spec_g_label in stages[0]
-
-        assert len(stages[1]) == 1
-        assert spec_d_label in stages[1]
-
-        assert len(stages[2]) == 1
-        assert spec_b_label in stages[2]
-
-        assert len(stages[3]) == 1
-        assert spec_a_label in stages[3]
 
 
 def test_ci_generate_with_env(ci_generate_test, tmp_path, mock_binary_index):
@@ -452,7 +398,10 @@ spack:
     assert "rebuild-index" not in yaml_contents
     assert "variables" in yaml_contents
     assert "SPACK_PIPELINE_TYPE" in yaml_contents["variables"]
-    assert yaml_contents["variables"]["SPACK_PIPELINE_TYPE"] == "spack_pull_request"
+    assert (
+        ci.common.PipelineType[yaml_contents["variables"]["SPACK_PIPELINE_TYPE"]]
+        == ci.common.PipelineType.PULL_REQUEST
+    )
 
 
 def test_ci_generate_with_external_pkg(ci_generate_test, tmp_path, monkeypatch):
@@ -1391,7 +1340,7 @@ spack:
             str(artifacts_root),
         )
 
-        job_name = ci.get_job_name(job_spec)
+        job_name = gitlab_generator.get_job_name(job_spec)
 
         with open(repro_dir / "repro.json", "w") as f:
             f.write(
@@ -1588,7 +1537,7 @@ spack:
     spec_a = Spec("pkg-a")
     spec_a.concretize()
 
-    return ci.get_job_name(spec_a)
+    return gitlab_generator.get_job_name(spec_a)
 
 
 def test_ci_dynamic_mapping_empty(
@@ -1604,7 +1553,7 @@ def test_ci_dynamic_mapping_empty(
     def fake_dyn_mapping_urlopener(*args, **kwargs):
         return BytesIO("{}".encode())
 
-    monkeypatch.setattr(ci, "_dyn_mapping_urlopener", fake_dyn_mapping_urlopener)
+    monkeypatch.setattr(ci.common, "_dyn_mapping_urlopener", fake_dyn_mapping_urlopener)
 
     _ = dynamic_mapping_setup(tmpdir)
     with tmpdir.as_cwd():
@@ -1633,7 +1582,7 @@ def test_ci_dynamic_mapping_full(
             ).encode()
         )
 
-    monkeypatch.setattr(ci, "_dyn_mapping_urlopener", fake_dyn_mapping_urlopener)
+    monkeypatch.setattr(ci.common, "_dyn_mapping_urlopener", fake_dyn_mapping_urlopener)
 
     label = dynamic_mapping_setup(tmpdir)
     with tmpdir.as_cwd():
@@ -1651,3 +1600,179 @@ def test_ci_dynamic_mapping_full(
                 assert job.get("variables", {}).get("MY_VAR") == "hello"
                 assert "ignored_field" not in job
                 assert "unallowed_field" not in job
+
+
+def test_ci_generate_unknown_generator(
+    ci_generate_test,
+    tmp_path,
+    mutable_mock_env_path,
+    install_mockery,
+    mock_packages,
+    ci_base_environment,
+):
+    """Ensure unrecognized ci targets are detected and the user
+    sees an intelligible and actionable message"""
+    src_mirror_url = tmp_path / "ci-src-mirror"
+    bin_mirror_url = tmp_path / "ci-bin-mirror"
+    spack_yaml_contents = f"""
+spack:
+  specs:
+    - archive-files
+  mirrors:
+    some-mirror: {src_mirror_url}
+    buildcache-destination: {bin_mirror_url}
+  ci:
+    target: unknown
+    pipeline-gen:
+    - submapping:
+      - match:
+          - archive-files
+        build-job:
+          tags:
+            - donotcare
+          image: donotcare
+"""
+    expect = "Spack CI module cannot generate a pipeline for format unknown"
+    with pytest.raises(ci.SpackCIError, match=expect):
+        ci_generate_test(spack_yaml_contents)
+
+
+def test_ci_generate_copy_only(
+    ci_generate_test,
+    tmp_path,
+    monkeypatch,
+    mutable_mock_env_path,
+    install_mockery,
+    mock_packages,
+    ci_base_environment,
+):
+    """Ensure the correct jobs are generated for a copy-only pipeline,
+    and verify that pipeline manifest is produced containing the right
+    number of entries."""
+    src_mirror_url = tmp_path / "ci-src-mirror"
+    bin_mirror_url = tmp_path / "ci-bin-mirror"
+    copy_mirror_url = tmp_path / "ci-copy-mirror"
+
+    monkeypatch.setenv("SPACK_PIPELINE_TYPE", "spack_copy_only")
+    monkeypatch.setenv("SPACK_COPY_BUILDCACHE", copy_mirror_url)
+
+    spack_yaml_contents = f"""
+spack:
+  specs:
+    - archive-files
+  mirrors:
+    buildcache-source:
+      fetch: {src_mirror_url}
+      push: {src_mirror_url}
+      source: False
+      binary: True
+    buildcache-destination:
+      fetch: {bin_mirror_url}
+      push: {bin_mirror_url}
+      source: False
+      binary: True
+  ci:
+    target: gitlab
+    pipeline-gen:
+    - submapping:
+      - match:
+          - archive-files
+        build-job:
+          tags:
+            - donotcare
+          image: donotcare
+"""
+    _, output_file, _ = ci_generate_test(spack_yaml_contents)
+
+    with open(output_file) as of:
+        pipeline_doc = syaml.load(of.read())
+
+    expected_keys = ["copy", "rebuild-index", "stages", "variables", "workflow"]
+    assert all([k in pipeline_doc for k in expected_keys])
+
+    # Make sure there are only two jobs and two stages
+    stages = pipeline_doc["stages"]
+    copy_stage = "copy"
+    rebuild_index_stage = "stage-rebuild-index"
+
+    assert len(stages) == 2
+    assert stages[0] == copy_stage
+    assert stages[1] == rebuild_index_stage
+
+    rebuild_index_job = pipeline_doc["rebuild-index"]
+    assert rebuild_index_job["stage"] == rebuild_index_stage
+
+    copy_job = pipeline_doc["copy"]
+    assert copy_job["stage"] == copy_stage
+
+    # Make sure a pipeline manifest was generated
+    output_directory = os.path.dirname(output_file)
+    assert "SPACK_ARTIFACTS_ROOT" in pipeline_doc["variables"]
+    artifacts_root = pipeline_doc["variables"]["SPACK_ARTIFACTS_ROOT"]
+    pipeline_manifest_path = os.path.join(
+        output_directory, artifacts_root, "specs_to_copy", "copy_rebuilt_specs.json"
+    )
+
+    assert os.path.exists(pipeline_manifest_path)
+    assert os.path.isfile(pipeline_manifest_path)
+
+    with open(pipeline_manifest_path) as fd:
+        manifest_data = json.load(fd)
+
+    with ev.read("test") as active_env:
+        active_env.concretize()
+        for s in active_env.all_specs():
+            assert s.dag_hash() in manifest_data
+
+
+@generator("unittestgenerator")
+def generate_unittest_pipeline(
+    pipeline: PipelineDag, spack_ci: SpackCIConfig, options: PipelineOptions
+):
+    """Define a custom pipeline generator for the target 'unittestgenerator'."""
+    output_file = options.output_file
+    assert output_file is not None
+    with open(output_file, "w") as fd:
+        fd.write("unittestpipeline\n")
+        for _, node in pipeline.traverse_nodes(direction="children"):
+            release_spec = node.spec
+            fd.write(f"  {release_spec.name}\n")
+
+
+def test_ci_generate_alternate_target(
+    ci_generate_test,
+    tmp_path,
+    mutable_mock_env_path,
+    install_mockery,
+    mock_packages,
+    ci_base_environment,
+):
+    """Ensure the above pipeline generator was correctly registered and
+    is used to generate a pipeline for the stack/config defined here."""
+    bin_mirror_url = tmp_path / "ci-bin-mirror"
+
+    spack_yaml_contents = f"""
+spack:
+  specs:
+    - archive-files
+    - externaltest
+  mirrors:
+    buildcache-destination: {bin_mirror_url}
+  ci:
+    target: unittestgenerator
+    pipeline-gen:
+    - submapping:
+      - match:
+          - archive-files
+        build-job:
+          tags:
+            - donotcare
+          image: donotcare
+"""
+    _, output_file, _ = ci_generate_test(spack_yaml_contents, "--no-prune-externals")
+
+    with open(output_file) as of:
+        pipeline_doc = of.read()
+
+    assert pipeline_doc.startswith("unittestpipeline")
+    assert "externaltest" in pipeline_doc
